refactor(layout): extract theme constants into themes.ts
Move AppTheme/THEMES/THEME_VARIANTS/ThemeVariant/ColorScheme/
ControlCenterCardsPerRow plus their type guards and splitAppTheme()
into a new components/layout/themes.ts. AppShell.tsx, SettingsScreenRoot.tsx
and new-shell/useSettingsState.ts now consume the shared module. Behavior
unchanged for both shells; the toggle is still live.

Side benefit: the new shell's useSettingsState now runs the legacy
combined-key migration too, so users upgrading from a pre-split build
keep their theme/color-scheme preference instead of resetting to
system+holaos.
